ghost-buster-ai/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.75.0",
    "fastapi[standard]>=0.115.0",
    "hrflow>=4.2.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
